requirements.txt
Flask>=3.0
Flask-Session>=0.8
Authlib>=1.3
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
requests
requests-oauthlib